ensuring sandbox is new
