+++ packages/destination-actions/src/destinations/tiktok-conversions/__tests__/index.test.ts
@@ -4,7 +4,7 @@ import Definition from '../index'
 import { Settings } from '../generated-types'
 
 const testDestination = createTestIntegration(Definition)
-const timestamp = '2024-01-08T13:52:50.212Z'
+const timestamp = '2021-09-2T15:21:15.449Z'
 const settings: Settings = {
   accessToken: 'test',
   pixelCode: 'test'
@@ -38,7 +38,7 @@ describe('Tiktok Conversions', () => {
         userId: 'testId123'
       })
 
-      nock('https://business-api.tiktok.com/open_api/v1.3/event/track').post('/').reply(200, {})
+      nock('https://business-api.tiktok.com/open_api/v1.3/pixel/track').post('/').reply(200, {})
       const responses = await testDestination.testAction('reportWebEvent', {
         event,
         settings,
@@ -47,48 +47,35 @@ describe('Tiktok Conversions', () => {
           event: 'InitiateCheckout'
         }
       })
-
       expect(responses.length).toBe(1)
       expect(responses[0].status).toBe(200)
       expect(responses[0].options.json).toMatchObject({
-        data: [
-          {
-            event: 'InitiateCheckout',
-            event_id: 'corey123',
-            event_time: 1704721970,
-            limited_data_use: false,
-            page: {
-              referrer: 'https://google.com/',
-              url: 'https://segment.com/'
-            },
-            properties: {
-              content_type: 'product',
-              contents: [],
-              currency: 'USD',
-              description: undefined,
-              order_id: undefined,
-              query: 'shoes',
-              shop_id: undefined,
-              value: 100
-            },
-            test_event_code: undefined,
-            user: {
-              email: ['eb9869a32b532840dd6aa714f7a872d21d6f650fc5aa933d9feefc64708969c7'],
-              external_id: ['481f202262e9c5ccc48d24e60798fadaa5f6ff1f8369f7ab927c04c3aa682a7f'],
-              ip: '0.0.0.0',
-              lead_id: undefined,
-              locale: undefined,
-              phone: ['910a625c4ba147b544e6bd2f267e130ae14c591b6ba9c25cb8573322dedbebd0'],
-              ttclid: '12345',
-              ttp: undefined,
-              user_agent:
-                'Mozilla/5.0 (iPhone; CPU iPhone OS 12_1_4 like Mac OS X) AppleWebKit/605.1.15 (KHTML, like Gecko) Mobile/16D57'
-            }
-          }
-        ],
-        event_source: 'web',
-        event_source_id: 'test',
-        partner_name: 'Segment'
+        pixel_code: 'test',
+        event: 'InitiateCheckout',
+        event_id: 'corey123',
+        timestamp: '2021-09-2T15:21:15.449Z',
+        context: {
+          user: {
+            external_id: '481f202262e9c5ccc48d24e60798fadaa5f6ff1f8369f7ab927c04c3aa682a7f',
+            phone_number: '910a625c4ba147b544e6bd2f267e130ae14c591b6ba9c25cb8573322dedbebd0',
+            email: 'eb9869a32b532840dd6aa714f7a872d21d6f650fc5aa933d9feefc64708969c7'
+          },
+          ad: {
+            callback: '12345'
+          },
+          page: {
+            url: 'https://segment.com/',
+            referrer: 'https://google.com/'
+          },
+          ip: '0.0.0.0',
+          user_agent:
+            'Mozilla/5.0 (iPhone; CPU iPhone OS 12_1_4 like Mac OS X) AppleWebKit/605.1.15 (KHTML, like Gecko) Mobile/16D57'
+        },
+        properties: {
+          currency: 'USD',
+          value: 100,
+          query: 'shoes'
+        }
       })
     })
 
@@ -105,16 +92,7 @@ describe('Tiktok Conversions', () => {
           currency: 'USD',
           value: 100,
           query: 'shoes',
-          products: [
-            {
-              price: 100,
-              quantity: 2,
-              category: 'Air Force One (Size S)',
-              product_id: 'abc123',
-              name: 'pname1',
-              brand: 'Brand X'
-            }
-          ]
+          products: [{ price: 100, quantity: 2, category: 'Air Force One (Size S)', product_id: 'abc123' }]
         },
         context: {
           page: {
@@ -128,7 +106,7 @@ describe('Tiktok Conversions', () => {
         userId: 'testId123'
       })
 
-      nock('https://business-api.tiktok.com/open_api/v1.3/event/track').post('/').reply(200, {})
+      nock('https://business-api.tiktok.com/open_api/v1.3/pixel/track').post('/').reply(200, {})
       const responses = await testDestination.testAction('reportWebEvent', {
         event,
         settings,
@@ -145,17 +123,11 @@ describe('Tiktok Conversions', () => {
                 quantity: {
                   '@path': '$.quantity'
                 },
-                content_category: {
+                content_type: {
                   '@path': '$.category'
                 },
                 content_id: {
                   '@path': '$.product_id'
-                },
-                content_name: {
-                  '@path': '$.name'
-                },
-                brand: {
-                  '@path': '$.brand'
                 }
               }
             ]
@@ -165,55 +137,9 @@ describe('Tiktok Conversions', () => {
 
       expect(responses.length).toBe(1)
       expect(responses[0].status).toBe(200)
-      expect(responses[0].options.json).toMatchObject({
-        data: [
-          {
-            event: 'InitiateCheckout',
-            event_id: 'corey123',
-            event_time: 1704721970,
-            limited_data_use: false,
-            page: {
-              referrer: 'https://google.com/',
-              url: 'https://segment.com/'
-            },
-            properties: {
-              content_type: 'product',
-              contents: [
-                {
-                  price: 100,
-                  quantity: 2,
-                  content_id: 'abc123',
-                  content_category: 'Air Force One (Size S)',
-                  content_name: 'pname1',
-                  brand: 'Brand X'
-                }
-              ],
-              currency: 'USD',
-              description: undefined,
-              order_id: undefined,
-              query: 'shoes',
-              shop_id: undefined,
-              value: 100
-            },
-            test_event_code: undefined,
-            user: {
-              email: ['eb9869a32b532840dd6aa714f7a872d21d6f650fc5aa933d9feefc64708969c7'],
-              external_id: ['481f202262e9c5ccc48d24e60798fadaa5f6ff1f8369f7ab927c04c3aa682a7f'],
-              ip: '0.0.0.0',
-              lead_id: undefined,
-              locale: undefined,
-              phone: ['910a625c4ba147b544e6bd2f267e130ae14c591b6ba9c25cb8573322dedbebd0'],
-              ttclid: '12345',
-              ttp: undefined,
-              user_agent:
-                'Mozilla/5.0 (iPhone; CPU iPhone OS 12_1_4 like Mac OS X) AppleWebKit/605.1.15 (KHTML, like Gecko) Mobile/16D57'
-            }
-          }
-        ],
-        event_source: 'web',
-        event_source_id: 'test',
-        partner_name: 'Segment'
-      })
+      expect(responses[0].options.body).toMatchInlineSnapshot(
+        `"{\\"pixel_code\\":\\"test\\",\\"event\\":\\"InitiateCheckout\\",\\"event_id\\":\\"corey123\\",\\"timestamp\\":\\"2021-09-2T15:21:15.449Z\\",\\"context\\":{\\"user\\":{\\"external_id\\":\\"481f202262e9c5ccc48d24e60798fadaa5f6ff1f8369f7ab927c04c3aa682a7f\\",\\"phone_number\\":\\"910a625c4ba147b544e6bd2f267e130ae14c591b6ba9c25cb8573322dedbebd0\\",\\"email\\":\\"eb9869a32b532840dd6aa714f7a872d21d6f650fc5aa933d9feefc64708969c7\\"},\\"ad\\":{\\"callback\\":\\"12345\\"},\\"page\\":{\\"url\\":\\"https://segment.com/\\",\\"referrer\\":\\"https://google.com/\\"},\\"ip\\":\\"0.0.0.0\\",\\"user_agent\\":\\"Mozilla/5.0 (iPhone; CPU iPhone OS 12_1_4 like Mac OS X) AppleWebKit/605.1.15 (KHTML, like Gecko) Mobile/16D57\\"},\\"properties\\":{\\"contents\\":[{\\"price\\":100,\\"quantity\\":2,\\"content_type\\":\\"Air Force One (Size S)\\",\\"content_id\\":\\"abc123\\"}],\\"currency\\":\\"USD\\",\\"value\\":100,\\"query\\":\\"shoes\\"},\\"partner_name\\":\\"Segment\\"}"`
+      )
     })
 
     it('should coerce properties into the contents array', async () => {
@@ -232,9 +158,7 @@ describe('Tiktok Conversions', () => {
           price: 100,
           quantity: 2,
           category: 'Air Force One (Size S)',
-          product_id: 'abc123',
-          name: 'pname1',
-          brand: 'Brand X'
+          product_id: 'abc123'
         },
         context: {
           page: {
@@ -248,7 +172,7 @@ describe('Tiktok Conversions', () => {
         userId: 'testId123'
       })
 
-      nock('https://business-api.tiktok.com/open_api/v1.3/event/track').post('/').reply(200, {})
+      nock('https://business-api.tiktok.com/open_api/v1.3/pixel/track').post('/').reply(200, {})
       const responses = await testDestination.testAction('reportWebEvent', {
         event,
         settings,
@@ -265,17 +189,11 @@ describe('Tiktok Conversions', () => {
                 quantity: {
                   '@path': '$.quantity'
                 },
-                content_category: {
+                content_type: {
                   '@path': '$.category'
                 },
                 content_id: {
                   '@path': '$.product_id'
-                },
-                content_name: {
-                  '@path': '$.name'
-                },
-                brand: {
-                  '@path': '$.brand'
                 }
               }
             ]
@@ -285,55 +203,9 @@ describe('Tiktok Conversions', () => {
 
       expect(responses.length).toBe(1)
       expect(responses[0].status).toBe(200)
-      expect(responses[0].options.json).toMatchObject({
-        data: [
-          {
-            event: 'AddToCart',
-            event_id: 'corey123',
-            event_time: 1704721970,
-            limited_data_use: false,
-            page: {
-              referrer: 'https://google.com/',
-              url: 'https://segment.com/'
-            },
-            properties: {
-              content_type: 'product',
-              contents: [
-                {
-                  price: 100,
-                  quantity: 2,
-                  content_id: 'abc123',
-                  content_category: 'Air Force One (Size S)',
-                  content_name: 'pname1',
-                  brand: 'Brand X'
-                }
-              ],
-              currency: 'USD',
-              description: undefined,
-              order_id: undefined,
-              query: 'shoes',
-              shop_id: undefined,
-              value: 100
-            },
-            test_event_code: undefined,
-            user: {
-              email: ['eb9869a32b532840dd6aa714f7a872d21d6f650fc5aa933d9feefc64708969c7'],
-              external_id: ['481f202262e9c5ccc48d24e60798fadaa5f6ff1f8369f7ab927c04c3aa682a7f'],
-              ip: '0.0.0.0',
-              lead_id: undefined,
-              locale: undefined,
-              phone: ['910a625c4ba147b544e6bd2f267e130ae14c591b6ba9c25cb8573322dedbebd0'],
-              ttclid: '12345',
-              ttp: undefined,
-              user_agent:
-                'Mozilla/5.0 (iPhone; CPU iPhone OS 12_1_4 like Mac OS X) AppleWebKit/605.1.15 (KHTML, like Gecko) Mobile/16D57'
-            }
-          }
-        ],
-        event_source: 'web',
-        event_source_id: 'test',
-        partner_name: 'Segment'
-      })
+      expect(responses[0].options.body).toMatchInlineSnapshot(
+        `"{\\"pixel_code\\":\\"test\\",\\"event\\":\\"AddToCart\\",\\"event_id\\":\\"corey123\\",\\"timestamp\\":\\"2021-09-2T15:21:15.449Z\\",\\"context\\":{\\"user\\":{\\"external_id\\":\\"481f202262e9c5ccc48d24e60798fadaa5f6ff1f8369f7ab927c04c3aa682a7f\\",\\"phone_number\\":\\"910a625c4ba147b544e6bd2f267e130ae14c591b6ba9c25cb8573322dedbebd0\\",\\"email\\":\\"eb9869a32b532840dd6aa714f7a872d21d6f650fc5aa933d9feefc64708969c7\\"},\\"ad\\":{\\"callback\\":\\"12345\\"},\\"page\\":{\\"url\\":\\"https://segment.com/\\",\\"referrer\\":\\"https://google.com/\\"},\\"ip\\":\\"0.0.0.0\\",\\"user_agent\\":\\"Mozilla/5.0 (iPhone; CPU iPhone OS 12_1_4 like Mac OS X) AppleWebKit/605.1.15 (KHTML, like Gecko) Mobile/16D57\\"},\\"properties\\":{\\"contents\\":[{\\"price\\":100,\\"quantity\\":2,\\"content_type\\":\\"Air Force One (Size S)\\",\\"content_id\\":\\"abc123\\"}],\\"currency\\":\\"USD\\",\\"value\\":100,\\"query\\":\\"shoes\\"},\\"partner_name\\":\\"Segment\\"}"`
+      )
     })
 
     it('should parse context.page.url ttclid if properties.ttclid not available', async () => {
@@ -351,9 +223,7 @@ describe('Tiktok Conversions', () => {
           price: 100,
           quantity: 2,
           category: 'Air Force One (Size S)',
-          product_id: 'abc123',
-          name: 'pname1',
-          brand: 'Brand X'
+          product_id: 'abc123'
         },
         context: {
           page: {
@@ -367,7 +237,7 @@ describe('Tiktok Conversions', () => {
         userId: 'testId123'
       })
 
-      nock('https://business-api.tiktok.com/open_api/v1.3/event/track').post('/').reply(200, {})
+      nock('https://business-api.tiktok.com/open_api/v1.3/pixel/track').post('/').reply(200, {})
       const responses = await testDestination.testAction('reportWebEvent', {
         event,
         settings,
@@ -384,17 +254,11 @@ describe('Tiktok Conversions', () => {
                 quantity: {
                   '@path': '$.quantity'
                 },
-                content_category: {
+                content_type: {
                   '@path': '$.category'
                 },
                 content_id: {
                   '@path': '$.product_id'
-                },
-                content_name: {
-                  '@path': '$.name'
-                },
-                brand: {
-                  '@path': '$.brand'
                 }
               }
             ]
@@ -404,55 +268,9 @@ describe('Tiktok Conversions', () => {
 
       expect(responses.length).toBe(1)
       expect(responses[0].status).toBe(200)
-      expect(responses[0].options.json).toMatchObject({
-        data: [
-          {
-            event: 'AddToCart',
-            event_id: 'corey123',
-            event_time: 1704721970,
-            limited_data_use: false,
-            page: {
-              referrer: 'https://google.com/',
-              url: 'http://demo.mywebsite.com?a=b&ttclid=123ATXSfe'
-            },
-            properties: {
-              content_type: 'product',
-              contents: [
-                {
-                  price: 100,
-                  quantity: 2,
-                  content_id: 'abc123',
-                  content_category: 'Air Force One (Size S)',
-                  content_name: 'pname1',
-                  brand: 'Brand X'
-                }
-              ],
-              currency: 'USD',
-              description: undefined,
-              order_id: undefined,
-              query: 'shoes',
-              shop_id: undefined,
-              value: 100
-            },
-            test_event_code: undefined,
-            user: {
-              email: ['eb9869a32b532840dd6aa714f7a872d21d6f650fc5aa933d9feefc64708969c7'],
-              external_id: ['481f202262e9c5ccc48d24e60798fadaa5f6ff1f8369f7ab927c04c3aa682a7f'],
-              ip: '0.0.0.0',
-              lead_id: undefined,
-              locale: undefined,
-              phone: ['910a625c4ba147b544e6bd2f267e130ae14c591b6ba9c25cb8573322dedbebd0'],
-              ttclid: '123ATXSfe',
-              ttp: undefined,
-              user_agent:
-                'Mozilla/5.0 (iPhone; CPU iPhone OS 12_1_4 like Mac OS X) AppleWebKit/605.1.15 (KHTML, like Gecko) Mobile/16D57'
-            }
-          }
-        ],
-        event_source: 'web',
-        event_source_id: 'test',
-        partner_name: 'Segment'
-      })
+      expect(responses[0].options.body).toMatchInlineSnapshot(
+        `"{\\"pixel_code\\":\\"test\\",\\"event\\":\\"AddToCart\\",\\"event_id\\":\\"corey123\\",\\"timestamp\\":\\"2021-09-2T15:21:15.449Z\\",\\"context\\":{\\"user\\":{\\"external_id\\":\\"481f202262e9c5ccc48d24e60798fadaa5f6ff1f8369f7ab927c04c3aa682a7f\\",\\"phone_number\\":\\"910a625c4ba147b544e6bd2f267e130ae14c591b6ba9c25cb8573322dedbebd0\\",\\"email\\":\\"eb9869a32b532840dd6aa714f7a872d21d6f650fc5aa933d9feefc64708969c7\\"},\\"ad\\":{\\"callback\\":\\"123ATXSfe\\"},\\"page\\":{\\"url\\":\\"http://demo.mywebsite.com?a=b&ttclid=123ATXSfe\\",\\"referrer\\":\\"https://google.com/\\"},\\"ip\\":\\"0.0.0.0\\",\\"user_agent\\":\\"Mozilla/5.0 (iPhone; CPU iPhone OS 12_1_4 like Mac OS X) AppleWebKit/605.1.15 (KHTML, like Gecko) Mobile/16D57\\"},\\"properties\\":{\\"contents\\":[{\\"price\\":100,\\"quantity\\":2,\\"content_type\\":\\"Air Force One (Size S)\\",\\"content_id\\":\\"abc123\\"}],\\"currency\\":\\"USD\\",\\"value\\":100,\\"query\\":\\"shoes\\"},\\"partner_name\\":\\"Segment\\"}"`
+      )
     })
 
     it('should send a successful lead_event event to reportWebEvent', async () => {
@@ -471,9 +289,7 @@ describe('Tiktok Conversions', () => {
           price: 100,
           quantity: 2,
           category: 'Air Force One (Size S)',
-          product_id: 'abc123',
-          name: 'pname1',
-          brand: 'Brand X'
+          product_id: 'abc123'
         },
         context: {
           page: {
@@ -487,7 +303,7 @@ describe('Tiktok Conversions', () => {
         userId: 'testId123'
       })
 
-      nock('https://business-api.tiktok.com/open_api/v1.3/event/track').post('/').reply(200, {})
+      nock('https://business-api.tiktok.com/open_api/v1.3/pixel/track').post('/').reply(200, {})
       const responses = await testDestination.testAction('reportWebEvent', {
         event,
         settings,
@@ -504,17 +320,11 @@ describe('Tiktok Conversions', () => {
                 quantity: {
                   '@path': '$.quantity'
                 },
-                content_category: {
+                content_type: {
                   '@path': '$.category'
                 },
                 content_id: {
                   '@path': '$.product_id'
-                },
-                content_name: {
-                  '@path': '$.name'
-                },
-                brand: {
-                  '@path': '$.brand'
                 }
               }
             ]
@@ -524,55 +334,9 @@ describe('Tiktok Conversions', () => {
 
       expect(responses.length).toBe(1)
       expect(responses[0].status).toBe(200)
-      expect(responses[0].options.json).toMatchObject({
-        data: [
-          {
-            event: 'lead_event',
-            event_id: 'corey123',
-            event_time: 1704721970,
-            limited_data_use: false,
-            page: {
-              referrer: 'https://google.com/',
-              url: 'http://demo.mywebsite.com?a=b&ttclid=123ATXSfe'
-            },
-            properties: {
-              content_type: 'product',
-              contents: [
-                {
-                  price: 100,
-                  quantity: 2,
-                  content_id: 'abc123',
-                  content_category: 'Air Force One (Size S)',
-                  content_name: 'pname1',
-                  brand: 'Brand X'
-                }
-              ],
-              currency: 'USD',
-              description: undefined,
-              order_id: undefined,
-              query: 'shoes',
-              shop_id: undefined,
-              value: 100
-            },
-            test_event_code: undefined,
-            user: {
-              email: ['eb9869a32b532840dd6aa714f7a872d21d6f650fc5aa933d9feefc64708969c7'],
-              external_id: ['481f202262e9c5ccc48d24e60798fadaa5f6ff1f8369f7ab927c04c3aa682a7f'],
-              ip: '0.0.0.0',
-              lead_id: '2229012621312',
-              locale: undefined,
-              phone: ['910a625c4ba147b544e6bd2f267e130ae14c591b6ba9c25cb8573322dedbebd0'],
-              ttclid: '123ATXSfe',
-              ttp: undefined,
-              user_agent:
-                'Mozilla/5.0 (iPhone; CPU iPhone OS 12_1_4 like Mac OS X) AppleWebKit/605.1.15 (KHTML, like Gecko) Mobile/16D57'
-            }
-          }
-        ],
-        event_source: 'web',
-        event_source_id: 'test',
-        partner_name: 'Segment'
-      })
+      expect(responses[0].options.body).toMatchInlineSnapshot(
+        `"{\\"pixel_code\\":\\"test\\",\\"event\\":\\"lead_event\\",\\"event_id\\":\\"corey123\\",\\"timestamp\\":\\"2021-09-2T15:21:15.449Z\\",\\"context\\":{\\"user\\":{\\"external_id\\":\\"481f202262e9c5ccc48d24e60798fadaa5f6ff1f8369f7ab927c04c3aa682a7f\\",\\"phone_number\\":\\"910a625c4ba147b544e6bd2f267e130ae14c591b6ba9c25cb8573322dedbebd0\\",\\"email\\":\\"eb9869a32b532840dd6aa714f7a872d21d6f650fc5aa933d9feefc64708969c7\\",\\"lead_id\\":\\"2229012621312\\"},\\"ad\\":{\\"callback\\":\\"123ATXSfe\\"},\\"page\\":{\\"url\\":\\"http://demo.mywebsite.com?a=b&ttclid=123ATXSfe\\",\\"referrer\\":\\"https://google.com/\\"},\\"ip\\":\\"0.0.0.0\\",\\"user_agent\\":\\"Mozilla/5.0 (iPhone; CPU iPhone OS 12_1_4 like Mac OS X) AppleWebKit/605.1.15 (KHTML, like Gecko) Mobile/16D57\\"},\\"properties\\":{\\"contents\\":[{\\"price\\":100,\\"quantity\\":2,\\"content_type\\":\\"Air Force One (Size S)\\",\\"content_id\\":\\"abc123\\"}],\\"currency\\":\\"USD\\",\\"value\\":100,\\"query\\":\\"shoes\\"},\\"partner_name\\":\\"Segment\\"}"`
+      )
     })
 
     it('should send test_event_code if present in mapping', async () => {
@@ -591,9 +355,7 @@ describe('Tiktok Conversions', () => {
           price: 100,
           quantity: 2,
           category: 'Air Force One (Size S)',
-          product_id: 'abc123',
-          name: 'pname1',
-          brand: 'Brand X'
+          product_id: 'abc123'
         },
         context: {
           page: {
@@ -607,7 +369,7 @@ describe('Tiktok Conversions', () => {
         userId: 'testId123'
       })
 
-      nock('https://business-api.tiktok.com/open_api/v1.3/event/track').post('/').reply(200, {})
+      nock('https://business-api.tiktok.com/open_api/v1.3/pixel/track').post('/').reply(200, {})
       const responses = await testDestination.testAction('reportWebEvent', {
         event,
         settings,
@@ -625,17 +387,11 @@ describe('Tiktok Conversions', () => {
                 quantity: {
                   '@path': '$.quantity'
                 },
-                content_category: {
+                content_type: {
                   '@path': '$.category'
                 },
                 content_id: {
                   '@path': '$.product_id'
-                },
-                content_name: {
-                  '@path': '$.name'
-                },
-                brand: {
-                  '@path': '$.brand'
                 }
               }
             ]
@@ -645,55 +401,9 @@ describe('Tiktok Conversions', () => {
 
       expect(responses.length).toBe(1)
       expect(responses[0].status).toBe(200)
-      expect(responses[0].options.json).toMatchObject({
-        data: [
-          {
-            event: 'AddToCart',
-            event_id: 'corey123',
-            event_time: 1704721970,
-            limited_data_use: false,
-            page: {
-              referrer: 'https://google.com/',
-              url: 'https://segment.com/'
-            },
-            properties: {
-              content_type: 'product',
-              contents: [
-                {
-                  price: 100,
-                  quantity: 2,
-                  content_id: 'abc123',
-                  content_category: 'Air Force One (Size S)',
-                  content_name: 'pname1',
-                  brand: 'Brand X'
-                }
-              ],
-              currency: 'USD',
-              description: undefined,
-              order_id: undefined,
-              query: 'shoes',
-              shop_id: undefined,
-              value: 100
-            },
-            test_event_code: 'TEST04030',
-            user: {
-              email: ['eb9869a32b532840dd6aa714f7a872d21d6f650fc5aa933d9feefc64708969c7'],
-              external_id: ['481f202262e9c5ccc48d24e60798fadaa5f6ff1f8369f7ab927c04c3aa682a7f'],
-              ip: '0.0.0.0',
-              lead_id: undefined,
-              locale: undefined,
-              phone: ['910a625c4ba147b544e6bd2f267e130ae14c591b6ba9c25cb8573322dedbebd0'],
-              ttclid: '12345',
-              ttp: undefined,
-              user_agent:
-                'Mozilla/5.0 (iPhone; CPU iPhone OS 12_1_4 like Mac OS X) AppleWebKit/605.1.15 (KHTML, like Gecko) Mobile/16D57'
-            }
-          }
-        ],
-        event_source: 'web',
-        event_source_id: 'test',
-        partner_name: 'Segment'
-      })
+      expect(responses[0].options.body).toMatchInlineSnapshot(
+        `"{\\"pixel_code\\":\\"test\\",\\"event\\":\\"AddToCart\\",\\"event_id\\":\\"corey123\\",\\"timestamp\\":\\"2021-09-2T15:21:15.449Z\\",\\"test_event_code\\":\\"TEST04030\\",\\"context\\":{\\"user\\":{\\"external_id\\":\\"481f202262e9c5ccc48d24e60798fadaa5f6ff1f8369f7ab927c04c3aa682a7f\\",\\"phone_number\\":\\"910a625c4ba147b544e6bd2f267e130ae14c591b6ba9c25cb8573322dedbebd0\\",\\"email\\":\\"eb9869a32b532840dd6aa714f7a872d21d6f650fc5aa933d9feefc64708969c7\\"},\\"ad\\":{\\"callback\\":\\"12345\\"},\\"page\\":{\\"url\\":\\"https://segment.com/\\",\\"referrer\\":\\"https://google.com/\\"},\\"ip\\":\\"0.0.0.0\\",\\"user_agent\\":\\"Mozilla/5.0 (iPhone; CPU iPhone OS 12_1_4 like Mac OS X) AppleWebKit/605.1.15 (KHTML, like Gecko) Mobile/16D57\\"},\\"properties\\":{\\"contents\\":[{\\"price\\":100,\\"quantity\\":2,\\"content_type\\":\\"Air Force One (Size S)\\",\\"content_id\\":\\"abc123\\"}],\\"currency\\":\\"USD\\",\\"value\\":100,\\"query\\":\\"shoes\\"},\\"partner_name\\":\\"Segment\\"}"`
+      )
     })
   })
 })
